hackathon day2/requirements.txt
langchain-core>=0.1.0
google-generativeai>=0.3.0
langchain>=0.1.0
faiss-cpu>=1.7.4
pypdf>=3.0.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.3
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0